% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{check_blob_exists}
\alias{check_blob_exists}
\title{Ensures that the filepath for the file to read exists}
\usage{
check_blob_exists(container, file, file_ext, info, path)
}
\arguments{
\item{container}{An Azure container object, as returned by 'get_container'}

\item{file}{The name of the file to be read, as a string. NB The file
extension does not need to be included (though it can be). The function
will error if multiple files are somehow matched.}

\item{file_ext}{The standard file extension for the file type, e.g. "json"}

\item{info}{Boolean. Whether to print user feedback about the file that is
being read. Useful for checking the function is doing what is expected, but
can be turned off with \code{FALSE}. Can be set persistently with the option
"azkit.info". If \code{NULL} then it will default to the value of
'rlang::is_interactive' (ie \code{TRUE} for interactive sessions).}

\item{path}{The path to the directory where \code{file} is located, as a string.
This must be the full path to the file location, as the function will not
search into subdirectories recursively. Set to \code{"/"} (the root of the
container) by default.}
}
\description{
Ensures that the filepath for the file to read exists
}
\keyword{internal}
