% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azkit_helpers.R
\name{check_container_class}
\alias{check_container_class}
\title{Check that a container looks like a real container}
\usage{
check_container_class(container)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}
}
\description{
Check that a container looks like a real container
}
