% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{check_envvar}
\alias{check_envvar}
\title{Check that an environment variable exists}
\usage{
check_envvar(x)
}
\arguments{
\item{x}{the \emph{name} of the environment variable to be found and checked}
}
\value{
the value of the environment variable \code{x}
}
\description{
The function prints a helpful error if the variable is not found, else
it returns the value of \code{Sys.getenv(x)}
}
