% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azkit_helpers.R
\name{check_nzchar}
\alias{check_nzchar}
\title{Check if a supplied non-NULL value is a string with >0 characters}
\usage{
check_nzchar(x, message, pf = parent.frame())
}
\arguments{
\item{x}{The object to be checked}

\item{message}{A custom error message, as a string. Will be shown to the
user if the check does not pass. Can include \code{glue} variables and \code{{cli}}
semantic markup. Variable values will be searched for in the environment of
the caller function (not in the environment of \code{check_nzchar()}). This
makes it easier to include informative values in the message.}

\item{pf}{Set as \code{\link[=parent.frame]{parent.frame()}} so variables in the caller environment can
be used in the custom error message.}
}
\description{
Will error if x is equal to \code{""}, or if it is otherwise missing or invalid.
With the exception that if x is NULL, then NULL will be passed through.
}
\keyword{internal}
