% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_that}
\alias{check_that}
\title{An alternative to stopifnot/assert_that etc}
\usage{
check_that(x, predicate, message, pf = parent.frame())
}
\arguments{
\item{x}{The object to be checked}

\item{predicate}{The predicate function used to check \code{x}}

\item{message}{A custom error message, as a string. Will be shown to the
user if the predicate check does not succeed. Can include \code{glue}d variables
and \code{{cli}} semantic markup.}

\item{pf}{Set as \code{\link[=parent.frame]{parent.frame()}} so variables in the caller environment can
be used in the custom error message.}
}
\description{
If the predicate function is true of \code{x} then \code{x} is returned. Otherwise,
an error is thrown with a custom \code{message}.
}
\seealso{
\link{check_vec}
}
