% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azkit_helpers.R
\name{check_val}
\alias{check_val}
\title{An alternative to stopifnot/assert_that
This function makes it easy to use the \code{{purrr}} functions \code{every()},
\code{some()} and \code{none()} to handle scalar values and vectors alike, and supports
the seamless use of \code{glue} strings in the custom error message.}
\usage{
check_val(
  x,
  predicate,
  message,
  which = c("every", "some", "none"),
  pf = parent.frame()
)
}
\description{
An alternative to stopifnot/assert_that
This function makes it easy to use the \code{{purrr}} functions \code{every()},
\code{some()} and \code{none()} to handle scalar values and vectors alike, and supports
the seamless use of \code{glue} strings in the custom error message.
}
\keyword{internal}
