% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cst_error_msg}
\alias{cst_error_msg}
\title{A custom error message generator for the \code{check_scalar_type} function}
\usage{
cst_error_msg(text)
}
\arguments{
\item{text}{The main text of the error message}
}
\description{
A custom error message generator for the \code{check_scalar_type} function
}
