% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_token.R
\name{generate_resource}
\alias{generate_resource}
\title{Generate appropriate values for the \code{resource} parameter in \link{get_auth_token}}
\usage{
generate_resource(
  version = 1,
  url = "https://storage.azure.com",
  path = "/.default",
  authorise = TRUE,
  refresh = TRUE
)
}
\arguments{
\item{version}{numeric. The AAD version, either 1 or 2 (1 by default)}

\item{url}{The URL of the Azure resource host}

\item{path}{For v2, the path designating the access scope}

\item{authorise}{Boolean, whether to return a token with authorisation scope,
(TRUE, the default) or one that just provides authentication. You are
unlikely to want to turn this off}

\item{refresh}{Boolean, applies to v2 tokens only, whether to return a token
that has a refresh token also supplied.}
}
\value{
A scalar character, or (in most v2 situations) a character vector
}
\description{
A helper function to generate appropriate values. Ensure that the \code{version}
argument matches the \code{aad_version} argument to \link{get_auth_token}.
It's unlikely that you will ever want to set \code{authorise} to \code{FALSE} but it's
here as an option since \link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token} supports it. Similarly,
you are likely to want to keep \code{refresh} turned on (this argument has no
effect on v1 tokens, it only applies to v2).
}
