% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{get_azure_endpoint}
\alias{get_azure_endpoint}
\title{Return an Azure "blob_endpoint"}
\usage{
get_azure_endpoint(token = NULL, endpoint_url = NULL, ...)
}
\arguments{
\item{token}{An Azure authentication token. If left as \code{NULL}, a token
returned by \link{get_auth_token} will be used}

\item{endpoint_url}{An Azure endpoint URL. If left as \code{NULL}, the default,
the value of the environment variable "AZ_STORAGE_EP" will be used}

\item{...}{arguments to be passed through to \link{get_auth_token}, if a token is
not already supplied}
}
\value{
An Azure blob endpoint (object of class "blob_endpoint")
}
\description{
This function will return the endpoint specified in the environment variable
"AZ_STORAGE_EP" by default
}
\keyword{internal}
