% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_azure_token.R
\name{get_client_id}
\alias{get_client_id}
\title{Sub-routine for \code{get_azure_token()}
Mainly to tidy up the code a bit}
\usage{
get_client_id()
}
\value{
A string (the client ID)
}
\description{
Sub-routine for \code{get_azure_token()}
Mainly to tidy up the code a bit
}
