% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{get_container}
\alias{get_container}
\title{Get Azure storage container}
\usage{
get_container(
  container_name,
  endpoint_url = Sys.getenv("AZ_STORAGE_EP"),
  token = get_auth_token()
)
}
\arguments{
\item{container_name}{Name of the container as a string.}

\item{endpoint_url}{An Azure endpoint URL.}

\item{token}{An Azure authentication token, or a function that returns one.
Uses \link{get_auth_token} by default.}
}
\value{
An Azure blob container (list object of class "blob_container")
}
\description{
It may be helpful to set the environment variable "AZ_STORAGE_EP". This
can contain your usual Azure storage endpoint URL should you not wish to
pass it in explicitly to the function.
You may find it helpful to use \link{list_container_names} to get a list of
available container names.
}
