% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{get_default_endpoint}
\alias{get_default_endpoint}
\title{Return an Azure "blob_endpoint"}
\usage{
get_default_endpoint(token)
}
\arguments{
\item{token}{An Azure authentication token}
}
\value{
An Azure blob endpoint (object of class "blob_endpoint")
}
\description{
Return an Azure "blob_endpoint"
}
\keyword{internal}
