% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gregg}
\alias{gregg}
\title{grepl a glued regex}
\usage{
gregg(x, rx, ..., g = parent.frame())
}
\arguments{
\item{x}{A character vector to check}

\item{rx}{A string that after processing by \code{\link[glue:glue]{glue::glue_data()}} will be used
as a regex pattern in \code{\link[=grepl]{grepl()}}}

\item{...}{Arguments passed onto \code{\link[=grepl]{grepl()}}}

\item{g}{The parent frame of gregg, which must be passed through to glue_data}
}
\description{
Use \{glue\} expressions in grepl (and put the arguments the right way round)
(https://glue.tidyverse.org/articles/wrappers.html)
}
\keyword{internal}
