% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{list_container_names}
\alias{list_container_names}
\title{Return a list of container names that are found at the endpoint}
\usage{
list_container_names(
  endpoint_url = Sys.getenv("AZ_STORAGE_EP"),
  token = get_auth_token()
)
}
\arguments{
\item{endpoint_url}{An Azure endpoint URL.}

\item{token}{An Azure authentication token, or a function that returns one.
Uses \link{get_auth_token} by default.}
}
\value{
A character vector of all container names found
}
\description{
Return a list of container names that are found at the endpoint
}
