% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files in a container}
\usage{
list_files(container, dir = "", ext = "", recursive = FALSE)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{dir}{(optional) The directory of the container to list files within.
\code{""} (the root directory of the container) by default}

\item{ext}{(optional) A string giving the extension of a particular file type
to restrict the list to. No need to include the initial ".". The default,
\code{""}, means no filtering by file extension will be applied.}

\item{recursive}{logical: whether to list files recursively. Default \code{FALSE}}
}
\value{
A vector of file names, or an empty character vector if none found
}
\description{
Lists all files (recursively, if desired) found in a container within a
given directory (\code{dir}). The search can be restricted to files with a
specific extension.
}
\details{
The function does not support filtering by file name, only by file extension.

The returned file list (character vector) contains the full paths to the
files, ready to be passed perhaps to a \verb{read_azure_*} function, or filtered
further. If you just want the names of the files without the folder path,
use \link{basename} to extract these.
}
\examples{
\dontrun{
  list_files(get_container("example"), ext = "csv")
}
}
