% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_token.R
\name{match_cached_token}
\alias{match_cached_token}
\title{Find an already cached token that matches desired parameters
Sub-routine for \code{get_auth_token()}}
\usage{
match_cached_token(resource, tenant, aad_version)
}
\value{
A token from local cache, or NULL if none matches
}
\description{
Find an already cached token that matches desired parameters
Sub-routine for \code{get_auth_token()}
}
\keyword{internal}
