% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_csv}
\alias{read_azure_csv}
\title{Read a csv file from Azure storage}
\usage{
read_azure_csv(container, file, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{file}{string The path to the file to be read.}

\item{...}{optional arguments to be passed through to \link[readr:read_delim]{readr::read_delim}}
}
\value{
A tibble
}
\description{
Read a csv file from Azure storage
}
