% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_file}
\alias{read_azure_file}
\title{Read any file from Azure storage}
\usage{
read_azure_file(container, file, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{file}{string The path to the file to be read.}

\item{...}{optional arguments to be passed through to
\link[AzureStor:blob]{AzureStor::download_blob}}
}
\value{
A raw data stream
}
\description{
Read any file from Azure storage
}
