% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_json}
\alias{read_azure_json}
\title{Read a json file from Azure storage}
\usage{
read_azure_json(container, file, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{file}{string The path to the file to be read.}

\item{...}{optional arguments to be passed through to
\link[yyjsonr:read_json_raw]{yyjsonr::read_json_raw}}
}
\value{
A list
}
\description{
Read a json file from Azure storage
}
