% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_jsongz}
\alias{read_azure_jsongz}
\title{Read a json.gz file from Azure storage}
\usage{
read_azure_jsongz(container, file, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{file}{string The path to the file to be read.}

\item{...}{optional arguments to be passed through to
\link[yyjsonr:read_json_file]{yyjsonr::read_json_file}}
}
\value{
A list
}
\description{
Read a json.gz file from Azure storage
}
