% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_parquet}
\alias{read_azure_parquet}
\title{Read a parquet file from Azure storage}
\usage{
read_azure_parquet(container, file, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{file}{string The path to the file to be read.}

\item{...}{optional arguments to be passed through to \link[arrow:read_parquet]{arrow::read_parquet}}
}
\value{
A tibble
}
\description{
Read a parquet file from Azure storage
}
\examples{
\dontrun{
  read_azure_parquet(cont, "data/folder/path/1.parquet")
}
}
