% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_rds}
\alias{read_azure_rds}
\title{Read an rds file from Azure storage}
\usage{
read_azure_rds(container, file, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{file}{string The path to the file to be read.}

\item{...}{optional arguments to be passed through to
\link[AzureStor:storage_save]{AzureStor::storage_load_rds}. For example, a compression type (one of
c("unknown", "gzip", "bzip2", "xz", "zstd", "none")) can be provided using
the argument \code{type}, which will be passed on to \link{memDecompress} via
\link[AzureStor:storage_save]{AzureStor::storage_load_rds}.}
}
\value{
The data object that was stored in the rds file
}
\description{
Read an rds file from Azure storage
}
