% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_token.R
\name{refresh_token}
\alias{refresh_token}
\title{Use the token's internal refresh() method to refresh it}
\usage{
refresh_token(token)
}
\arguments{
\item{token}{An Azure authentication token}
}
\value{
An Azure authentication token
}
\description{
Use the token's internal refresh() method to refresh it
}
