% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_token.R
\name{try_token_from_vars}
\alias{try_token_from_vars}
\title{Get token via app and secret environment variables
Sub-routine for \code{get_auth_token()}}
\usage{
try_token_from_vars(get_token_fun, ...)
}
\value{
A list with elements \code{result} and \code{error}. If this method is
successful, the \code{result} element will contain a token.
}
\description{
Get token via app and secret environment variables
Sub-routine for \code{get_auth_token()}
}
\keyword{internal}
