% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_token.R
\name{get_client_id}
\alias{get_client_id}
\title{Sub-routine for \code{get_auth_token()}}
\usage{
get_client_id()
}
\value{
A string (the client ID)
}
\description{
Pulled out mainly to tidy up the main function code a bit
}
\keyword{internal}
