% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{get_container}
\alias{get_container}
\title{Get Azure storage container}
\usage{
get_container(container_name, ...)
}
\arguments{
\item{container_name}{string: name of the container}

\item{...}{arguments to be passed through to \code{get_azure_token()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
