% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{list_container_names}
\alias{list_container_names}
\title{Return a list of container names that are found at the default endpoint}
\usage{
list_container_names(token = NULL, ...)
}
\arguments{
\item{token}{An Azure authentication token}

\item{...}{arguments to be passed through to \code{\link[=get_auth_token]{get_auth_token()}}}
}
\value{
A character vector of all container names found
}
\description{
Return a list of container names that are found at the default endpoint
}
