% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files in a container}
\usage{
list_files(container, path = "", ext = "", recursive = TRUE)
}
\arguments{
\item{container}{An Azure container object, as returned by \link{get_container}}

\item{path}{(optional) subdirectory of the container to list files within.
\code{""} (the root folder of the container) by default}

\item{ext}{(optional) A string giving the extension of a particular file type
you want to restrict the list to. No need to include the initial ".". The
default, \code{""}, means no filtering by file extension will be applied. Can be
a regular expression.}

\item{recursive}{A Boolean value: whether to list files recursively. \code{TRUE}
by default}
}
\value{
A vector of file names, or an empty character vector if none found
}
\description{
Recursively (or not, if desired) lists all files found in a container. Search
can be restricted to a particular 'subdirectory' of the container, and/or
to files with a specific extension. The function assumes that all file names
end with a ".ext" extension of some sort.
}
\details{
The function does not support filtering by file name, only by file extension.

The returned file list (character vector) contains the full paths to the
files, ready to be passed perhaps to a \verb{read_azure_*} function, or further
filtered by you. If you just want the names of the files without the folder
path, use \code{\link[=basename]{basename()}} to extract these.
}
\examples{
\dontrun{
  list_files(get_container("example"), ext = "csv")
}
}
