% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_files.R
\name{read_azure_parquet}
\alias{read_azure_parquet}
\title{Read a parquet file from Azure storage}
\usage{
read_azure_parquet(container, file, path = "/", info = NULL, ...)
}
\arguments{
\item{container}{An Azure container object, as returned by \code{\link[=get_container]{get_container()}}}

\item{file}{The name of the file to be read, as a string. NB The file
extension does not need to be included (though it can be). The function
will error if multiple files are somehow matched.}

\item{path}{The path to the directory where \code{file} is located, as a string.
This must be the full path to the file location, as the function will not
search into subdirectories recursively. Set to \code{"/"} (the root of the
container) by default.}

\item{info}{Boolean. Whether to print user feedback about the file that is
being read. Useful for checking the function is doing what is expected, but
can be turned off with \code{FALSE}. Can be set persistently with the option
"azkit.info". If \code{NULL} then it will default to the value of
\code{\link[rlang:is_interactive]{rlang::is_interactive()}} (ie \code{TRUE} for interactive sessions).}

\item{...}{optional arguments to be passed through to \code{\link[arrow:read_parquet]{arrow::read_parquet()}}}
}
\value{
A tibble
}
\description{
Read a parquet file from Azure storage
}
\examples{
\dontrun{
  # if a full filepath is available then path can be ignored
  read_azure_parquet(cont, "data/folder/path/1.parquet")
  # you can provide a filename without the '.parquet' extension
  # if you wish to use partial file name matching then it is probably easier
  # to provide a 'path'
  read_azure_parquet(cont, "case_details", "storage/parquet/2025/06/29")
}
}
