% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_table.R
\name{read_azure_table}
\alias{read_azure_table}
\title{Read in data from an Azure table}
\usage{
read_azure_table(
  table_name,
  table_endpoint = Sys.getenv("AZ_TABLE_EP"),
  token = get_auth_token(),
  filter = NULL,
  select = NULL,
  top = NULL
)
}
\arguments{
\item{table_name}{Name of the table to be read.}

\item{table_endpoint}{An Azure table endpoint URL.}

\item{token}{An Azure authentication token, or a function that returns one.
Uses \link{get_auth_token} by default.}

\item{filter}{An OData filter string to filter the results.}

\item{select}{An OData select string to specify which properties to return.}

\item{top}{An integer specifying the maximum number of records to return.}
}
\value{
A tibble
}
\description{
Read in data from an Azure table
}
