% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_table.R
\name{read_azure_table}
\alias{read_azure_table}
\title{Read in data from an Azure table}
\usage{
read_azure_table(
  table_name,
  table_endpoint = Sys.getenv("AZ_TABLE_EP"),
  token = get_auth_token()
)
}
\arguments{
\item{table_name}{Name of the table to be read.}

\item{table_endpoint}{An Azure table endpoint URL.}

\item{token}{An Azure authentication token, or a function that returns one.
Uses \link{get_auth_token} by default.}
}
\value{
A tibble
}
\description{
Read in data from an Azure table
}
