% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure_table.R
\name{read_azure_table}
\alias{read_azure_table}
\title{Read in data from an Azure table}
\usage{
read_azure_table(table_name = NULL, table_endpoint = NULL, ...)
}
\arguments{
\item{table_name}{name of the table to be read. If left as \code{NULL},
the default, the function will look instead for a value stored in the
environment variable "AZ_TABLE_NAME"}

\item{table_endpoint}{URL of the Azure table endpoint. If left as \code{NULL},
the default, the function will look instead for a value stored in the
environment variable "AZ_TABLE_EP"}

\item{...}{parameters to be passed through to \link{get_auth_token}}
}
\value{
A tibble
}
\description{
Read in data from an Azure table
}
